Fix unsafe cast from uint16_t* to size_t*.
This cast is not safe on 32-bit systems.

Once we drop support for AVR we can replace uint16_t with uint32_t
everywhere and life will be great.

+++ controller/test/controller_data_packet_ser/controller_data_packet_ser_test.cpp
@@ -12,7 +12,7 @@ uint8_t tx_buffer[PACKET_LEN_MAX];
 
 TEST(SerDesTests, StatusSerialization) {
 
-  size_t encoded_data_length;
+  uint16_t encoded_data_length;
   uint64_t time = 1;
   float pressure = 2.2;
   float volume = 3.3;
